+++ src/de/x28hd/tool/Condensr.java
@@ -17,7 +17,7 @@ public class Condensr extends JApplet {
 	private void initApplication(String[] args) {
 		System.setProperty("com.apple.mrj.application.apple.menu.about.name", "MyTool");
 		try {
-			PresentationService ps = new PresentationService(false);
+			GraphPanelControler ps = new GraphPanelControler(false);
 			new Thread(ps).start();
 			if (args.length >0) ps.setFilename(args[0], 0);
 		} catch (Throwable e) {
