+++ src/de/x28hd/tool/MyTool.java
@@ -15,7 +15,7 @@ public class MyTool extends JApplet {
 	private void initApplication(String[] args) {
 		System.setProperty("com.apple.mrj.application.apple.menu.about.name", "MyTool");
 		try {
-			PresentationService ps = new PresentationService(false);
+			GraphPanelControler ps = new GraphPanelControler(false);
 			new Thread(ps).start();
 			if (args.length >0) ps.setFilename(args[0], 0);
 		} catch (Throwable e) {
